Extract MoveToEffect and MoveByEffect into separate files
